+++ components/communityPage/SliderScreen.tsx
